feat: moved animation wrappers to own directory
