+++ packages/coil-extension/src/popup/components/page-animation-wrappers/AnimateTipCompleteViewMessageWrapper.tsx
