+++ packages/coil-extension/src/popup/components/page-animation-wrappers/AnimateTippingOpacityWrapper.tsx
@@ -3,8 +3,10 @@
 import React from 'react'
 import { motion } from 'framer-motion'
 
-import { useRouter } from '../context/routerContext'
-import { ROUTES } from '../constants'
+import { useRouter } from '../../context/routerContext'
+import { ROUTES } from '../../constants'
+
+import { pageTransitionDuration } from './animation-constants'
 
 //
 // Component
@@ -14,36 +16,32 @@ import { ROUTES } from '../constants'
     This animation is only for primary navigation transitions. 
     This does not handle the sub route navigation transitions such as from TipView -> TipConfirmView -> TipCompleteView
 */
-export const AnimateTippingNavBarWrapper = (
+export const AnimateTippingOpacityWrapper = (
   props: React.PropsWithChildren<any>
 ) => {
   const { path, previousPath } = useRouter()
-  const transitionDuration = 0.1
 
   // not sure where to put this, using it to determine the direction
   const isSubRoute =
     path.includes('tipping') && previousPath.includes('tipping')
 
   // this is used to prevent animation when navigating back to TipView from TipCompleteView
-  const preventAnimation = previousPath === ROUTES.tippingComplete
+  const preventAnimation = path === ROUTES.tippingComplete
 
   const primaryVariants = {
     initial: {
-      opacity: 1,
-      y: preventAnimation ? 0 : 62
+      opacity: 0
     },
     enter: {
       opacity: 1,
-      y: 0,
       transition: {
-        duration: preventAnimation ? 0 : transitionDuration
+        duration: pageTransitionDuration
       }
     },
     exit: {
       opacity: 0,
-      y: 62,
       transition: {
-        duration: preventAnimation ? 0 : transitionDuration
+        duration: preventAnimation ? 0 : pageTransitionDuration
       }
     }
   }
@@ -52,7 +50,6 @@ export const AnimateTippingNavBarWrapper = (
   // the rules used in the variants need to be overridden in order to eliminate a jump if the animation should not run.
   const notAnimated = {
     opacity: 1,
-    y: 0,
     transition: {
       duration: 0
     }
