+++ packages/coil-extension/src/popup/components/page-animation-wrappers/animation-constants.ts
@@ -1,0 +1,1 @@
+export const pageTransitionDuration = 0.1
