+++ packages/coil-extension/src/popup/components/views/TipView.tsx
@@ -11,10 +11,10 @@ import { CtaButton } from '../CtaButton'
 import { useStore } from '../../context/storeContext'
 import { useHost } from '../../context/popupHostContext'
 import { UpdateTippingSettings } from '../../../types/commands'
-import { AnimatePageTransitionWrapper } from '../AnimatePageTransitionWrapper'
-import { AnimateTippingHeaderWrapper } from '../AnimateTippingHeaderWrapper'
-import { AnimateTippingNavBarWrapper } from '../AnimateTippingNavBarWrapper'
-import { AnimateTippingOpacityWrapper } from '../AnimateTippingOpacityWrapper'
+import { AnimatePageTransitionWrapper } from '../page-animation-wrappers/AnimatePageTransitionWrapper'
+import { AnimateTippingHeaderWrapper } from '../page-animation-wrappers/AnimateTippingHeaderWrapper'
+import { AnimateTippingNavBarWrapper } from '../page-animation-wrappers/AnimateTippingNavBarWrapper'
+import { AnimateTippingOpacityWrapper } from '../page-animation-wrappers/AnimateTippingOpacityWrapper'
 import { Header } from '../Header'
 import { NavBar } from '../NavBar'
